breakpoints to match media sizes and colocate (#144)
