--- src/lib/upload/image-manipulation.ts
@@ -1,61 +1,0 @@
-import sharp from 'sharp'
-
-const BREAKPOINTS = {
-  large: 1000,
-  medium: 750,
-  small: 500,
-  thumbnail: 250,
-} as const
-
-type Breakpoint = keyof typeof BREAKPOINTS
-
-const breakpointKeys = Object.keys(BREAKPOINTS) as Breakpoint[]
-
-const getMetadata = async (filepath: string) => {
-  return await sharp(filepath).metadata()
-}
-
-const generateImage = async (
-  filepath: string,
-  filename: string,
-  breakpoint: number,
-  key: Breakpoint
-) => {
-  const folderPath = filepath.slice(0, -filename.length)
-  const outputPath = `${folderPath}${key}_${filename}`
-
-  await sharp(filepath)
-    .resize(breakpoint, breakpoint, {
-      fit: 'inside',
-    })
-    .toFile(outputPath)
-
-  const appendUrl = key + 'Url'
-  return { [appendUrl]: outputPath.slice(0, 5) }
-}
-
-const breakpointSmaller = (breakpoint: number, width = 0, height = 0) => {
-  return breakpoint < width || breakpoint < height
-}
-
-export const generateResponsiveImages = async (
-  filepath: string,
-  filename: string
-) => {
-  const { width, height } = await getMetadata(filepath)
-
-  const result = Promise.all(
-    breakpointKeys.map((key) => {
-      const breakpoint = BREAKPOINTS[key]
-
-      if (breakpointSmaller(breakpoint, width, height)) {
-        return generateImage(filepath, filename, breakpoint, key)
-      }
-      return undefined
-    })
-  )
-
-  return (await result).reduce((result, currentObj) => {
-    return { ...result, ...currentObj }
-  }, {})
-}
